+++ java/core/clustercontroller/src/main/java/io/joynr/messaging/routing/CcMessageRouter.java
@@ -92,4 +92,9 @@ public class CcMessageRouter extends AbstractMessageRouter {
             super.route(message);
         }
     }
+
+    @Override
+    protected boolean shutdownScheduler() {
+        return true;
+    }
 }