+++ java/core/libjoynr/src/main/java/io/joynr/messaging/routing/LibJoynrMessageRouter.java
@@ -217,4 +217,9 @@ public class LibJoynrMessageRouter extends AbstractMessageRouter {
     public void setIncomingAddress(Address incomingAddress) {
         this.incomingAddress = incomingAddress;
     }
+
+    @Override
+    protected boolean shutdownScheduler() {
+        return true;
+    }
 }