+++ wiki/ReleaseNotes.md
@@ -1,3 +1,14 @@
+# joynr 0.27.1
+
+## API relevant changes
+
+## Other changes
+* **[JEE]** Fixed cleanup of thread pool when application is undeployed
+* **[C++]** Mosquitto Connection uses now internal Mosquitto Loop thread handling.
+* **[C++]** Cleaned up MessagingSettings. All settings which correspond to `seconds` time unit,
+  have now `seconds` suffix.
+
+
 # joynr 0.27.0
 
 ## API relevant changes